add printer & jian ce qi

+++ HARDWARE/Drivers/BSP/Components/powerSTEP01/master_uart_control_interface.c
@@ -1,9 +1,9 @@
 #include "master_uart_control_interface.h"
 
 
-uint8_t Master_PowerStep01_Transfer_Interface(Powerstep1_contorl_motor_command_t* master,u8 length_data,u8* len){
+uint8_t Master_PowerStep01_Transfer_Interface(Powerstep1_contorl_motor_command_t* master,int length_data,int* len){
 		uint8_t ret=0;
-		u8 times =1000;
+		int times =1000;
 		master->OverReceiveFlag[0]=OVER_UART_VALUE0;
 		master->OverReceiveFlag[1]=OVER_UART_VALUE1;
 	
@@ -30,8 +30,8 @@ uint8_t Master_PowerStep01_Transfer_Interface(Powerstep1_contorl_motor_command_t
 
 uint8_t master_powerStep01_move_command(move_type_t data){
 		uint8_t ret =0;
-		u8 len =0;
-		u8 times =100;
+		int len =0;
+		int times =100;
 		Powerstep1_contorl_motor_command_t master_motorCommand;
 		master_motorCommand.type = MOVE_TYPE;
 		master_motorCommand.CommandPowerStep1.move=data;
@@ -54,8 +54,8 @@ uint8_t master_powerStep01_move_command(move_type_t data){
 
 uint8_t master_powerStep01_power_command(power_type_t data){
 		uint8_t ret =0;
-		u8 len =0;
-		u8 times =100;
+		int len =0;
+		int times =100;
 		Powerstep1_contorl_motor_command_t master_motorCommand;
 		master_motorCommand.type = POWER_TYPE;
 		master_motorCommand.CommandPowerStep1.power=data;
@@ -77,8 +77,8 @@ uint8_t master_powerStep01_power_command(power_type_t data){
 
 uint8_t master_powerStep01_rest_pos_command(rest_pos_type_t data){
 		uint8_t ret =0;
-		u8 len =0;
-		u8 times =100;
+		int len =0;
+		int times =100;
 		Powerstep1_contorl_motor_command_t master_motorCommand;
 		master_motorCommand.type = REST_TYPE;
 		master_motorCommand.CommandPowerStep1.rest_post=data;
@@ -101,8 +101,8 @@ uint8_t master_powerStep01_rest_pos_command(rest_pos_type_t data){
 
 uint8_t master_powerStep01_send_command_and_wait_no_busy(send_command_and_wait_no_busy_type_t data){
 		uint8_t ret =0;
-		u8 len =0;
-		u8 times =100;
+		int len =0;
+		int times =100;
 		Powerstep1_contorl_motor_command_t master_motorCommand;
 		master_motorCommand.type = SEND_COMMAND_AND_WAIT_NOTBUSY_TYPE;
 		master_motorCommand.CommandPowerStep1.send_and_wait=data;
@@ -123,8 +123,8 @@ uint8_t master_powerStep01_send_command_and_wait_no_busy(send_command_and_wait_n
 
 uint8_t master_powerStep01_one_device_move(one_device_move_type_t data){
 		uint8_t ret =0;
-		u8 len =0;
-		u8 times =100;
+		int len =0;
+		int times =100;
 		Powerstep1_contorl_motor_command_t master_motorCommand;
 		master_motorCommand.type = ONE_DEVICE_MOVE_TYPE;
 		master_motorCommand.CommandPowerStep1.one_device_move=data;
@@ -146,8 +146,8 @@ uint8_t master_powerStep01_one_device_move(one_device_move_type_t data){
 
 uint8_t master_powerStep01_one_device_wait(one_device_wait_type_t data){
 		uint8_t ret =0;
-		u8 len =0;
-		u8 times =100;
+		int len =0;
+		int times =100;
 		Powerstep1_contorl_motor_command_t master_motorCommand;
 		master_motorCommand.type = ONE_DEVICE_WAIT_TYPE;
 		master_motorCommand.CommandPowerStep1.one_device_wait=data;
@@ -168,8 +168,8 @@ uint8_t master_powerStep01_one_device_wait(one_device_wait_type_t data){
 
 uint8_t master_powerStep01_one_device_get_pos(one_device_get_pos_type_t data,int32_t*pos){
 		uint8_t ret =0;
-		u8 len =0;
-		u8 times =100;
+		int len =0;
+		int times =100;
 		Powerstep1_contorl_motor_command_t master_motorCommand;
 		master_motorCommand.type = ONE_DEVICE_GET_POS_TYPE;
 		master_motorCommand.CommandPowerStep1.one_device_get_pos=data;
@@ -192,8 +192,8 @@ uint8_t master_powerStep01_one_device_get_pos(one_device_get_pos_type_t data,int
 
 uint8_t master_powerStep01_one_device_set_mark(one_device_set_mark_type_t data){
 		uint8_t ret =0;
-		u8 len =0;
-		u8 times =100;
+		int len =0;
+		int times =100;
 		Powerstep1_contorl_motor_command_t master_motorCommand;
 		master_motorCommand.type = ONE_DEVICE_SET_MARK_TYPE;
 		master_motorCommand.CommandPowerStep1.one_device_set_mark=data;
@@ -216,8 +216,8 @@ uint8_t master_powerStep01_one_device_set_mark(one_device_set_mark_type_t data){
 
 uint8_t master_powerStep01_get_para(get_para_type_t data,powerstep01_Registers_t*para){
 		uint8_t ret =0;
-		u8 len =0;
-		u8 times =100;
+		int len =0;
+		int times =100;
 		Powerstep1_contorl_motor_command_t master_motorCommand;
 		master_motorCommand.type = GET_PARAM_TYPE;
 		master_motorCommand.CommandPowerStep1.get_para=data;
@@ -240,8 +240,8 @@ uint8_t master_powerStep01_get_para(get_para_type_t data,powerstep01_Registers_t
 
 uint8_t master_powerStep01_set_para(set_para_type_t data){
 		uint8_t ret =0;
-		u8 len =0;
-		u8 times =100;
+		int len =0;
+		int times =100;
 		Powerstep1_contorl_motor_command_t master_motorCommand;
 		master_motorCommand.type = SET_PARAM_TYPE;
 		master_motorCommand.CommandPowerStep1.set_para=data;
@@ -263,8 +263,8 @@ uint8_t master_powerStep01_set_para(set_para_type_t data){
 
 uint8_t master_powerStep01_select_step_mode(select_step_mode_t data){
 		uint8_t ret =0;
-		u8 len =0;
-		u8 times =100;
+		int len =0;
+		int times =100;
 		Powerstep1_contorl_motor_command_t master_motorCommand;
 		master_motorCommand.type = SElECT_STEP_MODE_TYPE;
 		master_motorCommand.CommandPowerStep1.select_step_mode=data;
@@ -286,8 +286,8 @@ uint8_t master_powerStep01_select_step_mode(select_step_mode_t data){
 
 uint8_t master_get_light_sensor_level(get_light_sensor_level_t data,uint8_t*value){
 		uint8_t ret =0;
-		u8 len =0;
-		u8 times =100;
+		int len =0;
+		int times =100;
 		Powerstep1_contorl_motor_command_t master_motorCommand;
 		master_motorCommand.type = GET_LIGHT_LEVEL_TYPE;
 		master_motorCommand.CommandPowerStep1.get_light_sensor_level=data;
@@ -312,7 +312,7 @@ uint8_t master_get_light_sensor_level(get_light_sensor_level_t data,uint8_t*valu
 //cheminert_c52_type_t
 uint8_t master_cheminert_c52_c55(cheminert_c52_c55_type_t data,uint8_t*buf,uint8_t*buf_size){
 		uint8_t ret =0;
-		u8 len =0;
+		int len =0;
 		u8 size=0;
 		u8 times =100;
 		Powerstep1_contorl_motor_command_t master_motorCommand;
@@ -341,7 +341,7 @@ uint8_t master_cheminert_c52_c55(cheminert_c52_c55_type_t data,uint8_t*buf,uint8
 
 uint8_t master_pump_s100_interface(pump_s100_command_type_t data,pump_s100_reply_type_t* reply,PUMP_S100_REPLY_type_t* type){
 		uint8_t ret =0;
-		u8 len =0;
+		int len =0;
 		u8 size=0;
 		u8 times =100;
 		Powerstep1_contorl_motor_command_t master_motorCommand;
